requirements.txt
numpy
torch
torchvision
scikit-learn
scipy
ipython
ipdb
yapf
tensorboardX
omegaconf
tqdm
loguru
ruamel.yaml
matplotlib
seaborn